LF-4164 Refactor: Move forward and backward synch logic into hooks; clean up type workaround
